+++ src/nodes/entities/custom-emoji.ts
@@ -1,9 +1,7 @@
 // Telegram Abstract Syntax Tree (tgast)
 // https://github.com/tgast-ecosystem/tgast
 
-import type { Data, Parent } from '../../abstract.ts'
-import type { Style } from '../../registries/style.ts'
-import type { Text } from '../text.ts'
+import type { Data, Literal } from '../../abstract.ts'
 
 /**
  * Extra data for {@linkcode CustomEmoji} nodes.
@@ -13,7 +11,7 @@ export interface CustomEmojiData extends Data { }
 /**
  * Telegram custom emoji entity.
  */
-export interface CustomEmoji extends Parent {
+export interface CustomEmoji extends Literal {
   /**
    * Node type is `"custom_emoji"`.
    */
@@ -26,10 +24,6 @@ export interface CustomEmoji extends Parent {
    * [getCustomEmojiStickers]: https://core.telegram.org/bots/api#getcustomemojistickers
    */
   emojiId: string
-  /**
-   * Content of the custom emoji text.
-   */
-  children: Array<Style | Text>
   /**
    * Extra data.
    */